Creating rotations for combined divisions
A great deal of work was involved in this release and although not all
the pieces are perfect, this version provides for a complete
RotationsPane.java. Some IO issues may persist but are unlikely. A
revamped Factory.java allows for sorting multiple (>= 2) arrays at a
time based on a template array.

This has been a long week.