Culmination of combined division effort. Start of empty() methods.
This revision attempts to put behind some major headaches associated
with the combined division issues when entering finishes. Of interest,
the FinishListItem has been revamped such that it is no longer
subclasses regatta.Finish, even though it holds its own finish
object.

In other news, the empty() method was introduced to AbstractPane to
allow panes to decide whether or not to allow leaving the current
pane, i.e. to commit any pending changes if necessary. The default
action is to permit changes without doing anything.

Also, the DailySummaryPane has been finished. This include bring the
XML format of the files to match TechScore-web v2.0 with respect to
daily summaries. The DetailsPane instead displays some statistics
(pending). This needs some cosmetic fixes.

There is a known issue with opening regattas that have been saved
using combined scoring in that the finishes appear as incomplete. It
is not clear whether this is a problem with saving (unlikely) or with
the "normalizeFinishes(...)" method.