Migration from MAKE to ANT
This revision is a major change in the file structure of the entire
library, utilizing the scheme from Android development. Makefiles are
no longer part of the distribution. Instead, the ANT tool is
favored. Note that the documentation (README) does not yet reflect
these changes.

Resources such as images and CSS files have been moved under the 'res'
folder.

No major code changes are present in this revision.