Improvements associated with ML request for fixes.
This release moves the RP membership to the MembershipDatabase class
which still uses the temporary text files as the database. The
setRegatta method in AbstractPane has been de-finalized in order to
make sure that RPFormPane can properly dispose of the membership
database results when the regatta is set to null.

The RPFormPane can now also directly add new sailors to the database,
and the membership database, along with the file format, tracks which
sailors are new (and therefore editable) and which are not. It is
important to note that sailors are not EDITED using RPFormPane, they
can only be added. To edit existing sailors, the RPDatabasePane must
still be used.

The basic framework for the context help files has also been created,
although no documentation yet exists.